requirements.txt
azure-mgmt-resource
azure-identity
azure-mgmt-consumption
azure-mgmt-subscription
azure-mgmt-storage 
azure-mgmt-cognitiveservices 
azure-mgmt-search
azure-storage-blob